backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
openai==1.12.0